requirements.txt
Flask==3.0.0
Flask-FlatPages==0.8.2
Frozen-Flask==1.0.1
Markdown==3.5.1
python-frontmatter==1.1.0
PyYAML==6.0.1


